Revert "Refactor NDK helpers some more, add methods for SurfaceControl/Transa…"
This reverts commit a119a0dca83ec2818830fcc775058a1d3d84392a.
