+++ ci/licenses_golden/excluded_files
@@ -100,7 +100,6 @@
 ../../../flutter/fml/message_loop_task_queues_unittests.cc
 ../../../flutter/fml/message_loop_unittests.cc
 ../../../flutter/fml/paths_unittests.cc
-../../../flutter/fml/platform/android/ndk_helpers_unittests.cc
 ../../../flutter/fml/platform/darwin/cf_utils_unittests.mm
 ../../../flutter/fml/platform/darwin/scoped_nsobject_arc_unittests.mm
 ../../../flutter/fml/platform/darwin/scoped_nsobject_unittests.mm
