+++ shell/platform/android/image_external_texture_gl.cc
@@ -8,11 +8,11 @@
 #include <android/sensor.h>
 
 #include "flutter/common/graphics/texture.h"
-#include "flutter/fml/platform/android/ndk_helpers.h"
 #include "flutter/impeller/core/formats.h"
 #include "flutter/impeller/display_list/dl_image_impeller.h"
 #include "flutter/impeller/toolkit/egl/image.h"
 #include "flutter/impeller/toolkit/gles/texture.h"
+#include "flutter/shell/platform/android/ndk_helpers.h"
 #include "third_party/skia/include/core/SkAlphaType.h"
 #include "third_party/skia/include/core/SkColorType.h"
 #include "third_party/skia/include/gpu/ganesh/SkImageGanesh.h"
