+++ shell/platform/android/image_external_texture_gl.h
@@ -17,8 +17,8 @@
 #include "flutter/impeller/toolkit/egl/image.h"
 #include "flutter/impeller/toolkit/gles/texture.h"
 
-#include "flutter/fml/platform/android/ndk_helpers.h"
 #include "flutter/shell/platform/android/android_context_gl_skia.h"
+#include "flutter/shell/platform/android/ndk_helpers.h"
 
 namespace flutter {
