+++ shell/platform/android/image_lru.h
@@ -9,7 +9,7 @@
 #include <cstddef>
 
 #include "display_list/image/dl_image.h"
-#include "fml/platform/android/ndk_helpers.h"
+#include "shell/platform/android/ndk_helpers.h"
 
 namespace flutter {
