+++ shell/platform/android/library_loader.cc
@@ -3,9 +3,9 @@
 // found in the LICENSE file.
 
 #include "flutter/fml/platform/android/jni_util.h"
-#include "flutter/fml/platform/android/ndk_helpers.h"
 #include "flutter/shell/platform/android/android_image_generator.h"
 #include "flutter/shell/platform/android/flutter_main.h"
+#include "flutter/shell/platform/android/ndk_helpers.h"
 #include "flutter/shell/platform/android/platform_view_android.h"
 #include "flutter/shell/platform/android/vsync_waiter_android.h"
